requirements.txt
# TensorFlow CPU ligero (Render no soporta GPU)
tensorflow-cpu==2.13.0
keras==2.13.1

# Backend
fastapi==0.95.2
uvicorn==0.22.0

# Procesamiento de imágenes
pillow==10.0.0
pandas==2.1.1
numpy==1.24.3   # Numpy óptimo para TensorFlow 2.13.0

# Base de datos
mysql-connector-python==8.4.0

# Utilidades
joblib==1.3.2
python-multipart==0.0.6
gdown==4.7.1
